fix : shifted the ham menu to navbar
