+++ src/components/cssComponents/Card.js
@@ -1,5 +1,5 @@
 import React from "react";
-import App from "./../../App";
+import App from "../../Base";
 function Card() {
   return (
     <App
