+++ src/components/cssComponents/Navbar.js
@@ -1,5 +1,5 @@
 import React from "react";
-import App from "./../../App";
+import App from "../../Base";
 
 function Navbar() {
   return (
