+++ src/components/cssComponents/Slider.js
@@ -1,5 +1,5 @@
 import React from "react";
-import App from "./../../App";
+import App from "../../Base";
 
 function Slider() {
   return (
